docker_images/api/song_adder_controller/requirements.txt
fastapi==0.87.0
classy-fastapi==0.2.12
fastapi-utils==0.2.1
starlette==0.21.0
SQLAlchemy==1.4.42
python-jose==3.3.0
uvicorn==0.19.0
requests==2.25.1
requests-oauthlib==1.3.1
python-multipart==0.0.5
pika==1.3.0
pika-stubs==0.1.3
retry==0.9.2
redis==2.10.6